refactor admin components, sections and pages
